Split tests and example/integration tests
