+++ cdi2/src/test/java/io/cucumber/cdi2/ApplicationScopedBean.java
@@ -1,0 +1,8 @@
+package io.cucumber.cdi2;
+
+import javax.enterprise.context.ApplicationScoped;
+
+@ApplicationScoped
+public class ApplicationScopedBean {
+
+}
