+++ cdi2/src/test/java/io/cucumber/cdi2/UnmanagedBean.java
@@ -1,0 +1,5 @@
+package io.cucumber.cdi2;
+
+public class UnmanagedBean {
+
+}
