+++ cdi2/src/test/java/io/cucumber/cdi2/example/BellyStepDefinitions.java
@@ -1,0 +1,34 @@
+package io.cucumber.cdi2.example;
+
+import io.cucumber.cdi2.Belly;
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+
+import javax.enterprise.inject.Vetoed;
+import javax.inject.Inject;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+@Vetoed
+public class BellyStepDefinitions {
+
+    @Inject
+    private Belly belly;
+
+    @Given("I have {int} cukes in my belly")
+    public void haveCukes(int n) {
+        belly.setCukes(n);
+    }
+
+    @Given("I eat {int} more cukes")
+    public void addCukes(int n) {
+        belly.setCukes(belly.getCukes() + n);
+    }
+
+    @Then("there are {int} cukes in my belly")
+    public void checkCukes(int n) {
+        assertEquals(n, belly.getCukes());
+    }
+
+
+}
