+++ cdi2/src/test/java/io/cucumber/cdi2/example/RunCucumberTest.java
@@ -1,4 +1,4 @@
-package io.cucumber.cdi2;
+package io.cucumber.cdi2.example;
 
 import io.cucumber.junit.Cucumber;
 import org.junit.runner.RunWith;
