+++ cdi2/src/test/resources/META-INF/beans.xml
@@ -5,6 +5,6 @@
     <!-- Exclude the class to force it to be loaded as an unmanaged dependency -->
     <scan>
         <exclude
-            name="io.cucumber.cdi2.UnmanagedBellyStepDefinitions" />
+            name="io.cucumber.cdi2.UnmanagedBean" />
     </scan>
 </beans>
