+++ cdi2/src/test/resources/io/cucumber/cdi2/example/cukes.feature
@@ -6,9 +6,5 @@ Feature: Cukes
 
   Scenario: Eat some more cukes
     Given I have 6 cukes in my belly
-    Then there are 6 cukes in my belly
-
-  Scenario: Eat some unmanaged cukes
-    Given I have 4 unmanaged cukes in my belly
     And I eat 2 more cukes
-    Then there are 6 cukes in my belly
+    Then there are 8 cukes in my belly
